User code stops sys oled
Allow oled-clients to get the device in either exclusive or non-exclusive mode, with the former default. This allows the sys-oled menu to run in non-exclusive mode, detect when another client connects and stop rendering. Use the atexit lib to ensure that only when an oled-client exits and therefore is completely done with the device, it releases the lock on the device. Restart the pt-sys-oled service completely when the user is done with the oled, so that it re-initialises correctly