refactor: extract scroll logic to paginated tile
Add a PaginatedTile class which takes a list of pages, renders the
current page and handles scrolling between pages. Currently the state
and logic for this is split between the Menu and MenuTile classes. Menu
stores the pages and current page while MenuTile handles scrolling
and updating the internal state of its Menu.

Remove the Menu class and pass pages directly to MenuTile. Extracting
pagination related state means Menu would only exist to store pages.
Passing pages to MenuTile instead allows us to remove the Menu
abstraction entirely.

Step the scroll position in the PaginatedTile's scroll loop. MenuTile
relies on it's image being accessed to step the animation, this
means the speed of the animation depends on how the parent behaves.
Updating scroll position in the scroll loop eliminates this
interdependence and makes the logic more robust as a result.

Inherit from PaginatedTile in MenuTile. This allows MenuTile to focus on
Menu specific behaviour, such as moving into a child menu. The MenuTile
still handles starting actions which should be refactored out later.