feat: add scrollbar to PaginatedTile
Add scrollbar hotspot to PaginatedTile. While PaginatedTile is
inheriting from ViewportTile it is not possible to add a hotspot that
scrolls independently of the pages, inherit from Tile instead. In order
for PaginatedTile to render the scrollbar hotspot as well as the
ViewportTile the `get_preprocess_image` method needs to be overridden.
This is so the initial image can have the pages ViewportTile's image
pasted in.

Override the PaginatedTile's `__setattr__` method in order to keep the
`active` state of the pages ViewportTile in sync.

Pass Page classes instead of instances to PaginatedTile so they can be
instantiated with the correct size. The size must now take the scrollbar
into account.

Some pages have crop and/or position issues due to the area to render
into becoming smaller. Fix some pages such as the Battery, Overview and
CPU pages and leave the rest. This is because a right gutter will be
added soon, and they will need to be adjusted again at that point.

Fix bug in `pause_every` generator where if the pause_yield_interval is
0 there is a exception due to the modulo.