feat: organise pages into sub-menus (#209)
The existing organisation has two issues:
- the flat structure of the information pages doesn't scale well
- the settings menu is not discoverable

Organise the pages into sub-menus found on the main menu. Each sub-menu
is entered and exited using the circle and cancel buttons respectively.
Moving the pages into sub-menus allows for more pages to be added
without making the main menu overly long.

Move the settings menu into the main menu as it's own sub-menu. Adding
the settings menu as a sub-menu of the main menu means users will
discover it more easily by scrolling through the sub-menus.

Flatten the settings menu since it does not have a large number of
pages and keeps all the sub-menus flat.

The new structure makes the behaviour of the cancel button easier to
understand since it no longer has multiple purposes, it only moves users
out of a menu.

Copy the battery page into a new 'overview' page, this is just a
placeholder and will be updated to include extra information than just
the battery level.