fix: hotspot memory leaks
Since the Hotspot class's caching thread is long lived and uses self
there is a circular reference that prevents hotspots being garbage
collected.

The memory leak causes Hotspots to stay in memory for the lifetime of
the application. When navigating to and from child menus the number of
Hotspots in memory increases which impacts performance noticably over
time.

Add a cleanup method to the Hotspot class that stops the caching thread.
When the thread is stopped it no longer references self and the hotspot
can be collected.

Cleanup hotspots when their Tile removes them or the Tile itself is
garbage collected.