fix: battery page memory leak
The Battery class uses long lived threads under the hood for it's
callbacks, for example the `on_capacity_change` callback. When passing
callbacks to the Battery self is referenced which creates a cirular
reference between the Battery and the Page class.

The Battery class also seems to have internal behaviour that prevents it
being garbage collected at all.

When navigating to the SystemMenu and back multiple times multiple
instances of BatteryPage are created but never garbage collected. This
causes a noticable performance impact over time.

Add a cleanup method to the Page class and implement it in BatteryPage
to remove the callbacks from the Battery. Removing the callbacks removes
the circular references and allows BatteryPage to be garbage collected.
Call the Page's cleanup method when it's parent PaginatedTile is garbage
collected.

Move the Battery to the module level rather than the class level. Since
Battery is never garbage collected creating a new Battery instance every
time BatteryPage is created also causes a memory leak. This has the
added benefit of speeding up the time it takes to create the SystemMenu.