feat: add indicators for right buttons (#223)
Add a gutter to the right of the HUDTileGroup with indicators for the
buttons on the right hand side of the miniscreen.

The indicators show that the select button can navigate into a menu with
a right arrow, or can perform an action with a tick. The cancel button
has a left arrow next to it when it's possible to navigate out of a
sub-menu.

Change the indicators when the current page of the top most
PaginatedTile changes, or when navigating in or out of a sub-menu.

Add events for when the PaginatedTile's scrolling animation starts or
finishes. This allows the indicators to be changed at the end of the
scroll animation.

Update the overview, battery, cpu, action and network pages to fit the
new page size.

Remove the icons from the action pages since they are not resizable and
could not fit with the more important text easily.

Refactor the battery page so that it is easier to work with.