refactor: fix HUD giving children the wrong size
When adding the right gutter to the HUDTileGroup the size of child tiles
should have been reduced by the width of the gutter.

The gutter was manually accounted for in the child tiles and so this
does not change any user facing layout.

In the base network page the right margin had 10 extra pixels on it.

In the base action page the wrong column was actually passed to the
text, fixing this error actually caused the layout to be unchanged.