fix: run script not using local installation (#229)
`sudo python3 ./dev/run.py` is not using the local development
installation. Since it is not in the root directory sys.path does not
have the correct path to find the local installation. Move `run.py` to
the root directory and rename it `dev-start.py`. Update `dev/README.md`
for the new file.