feat: add boundary alignment to text hotspot (#233)
Align text horizonally to boundaries when `align` is set. Options match
Pillow's `align` options: "left", "right" and "center" where "left" is
the default.

Add `vertical_align` property that aligns text similarly to `align` but
with vertical boundaries. Options are "top", "bottom" and "center" where
"top" is the default.

Set `vertical_align` to "center" in the base action page. Since the
previous default was to center the text horizontally and vertically the
action page text needs to be manually set to "center"

Remove `xy` and `anchor` properties since they are now superceded by the
align properties. Anchor is always left-top "lt" so that text xy
position can be calculated based off align values, text size and hotspot
size.

Incorperate MiniscreenAssistant logic into the TextHotspot. This allows
us to have greater control on the behaviour of the TextHotspot without
having to consider the impact of breaking changes.

Refactor the logic in the assistant for wrapping text and add it as a
`create_wrapped_text` method.

Memoize the `get_text_size` and `create_wrapped_text` methods since they
are called frequently and are potentially expensive.