fix: overview starting with no capacity text
The overview page takes about a second to render capacity text next to
the battery. This seems to be because the text hotspot doesn't
initialise it's cached image and the initial render of the page happens
before it's first render cycle.

Render the cached image when initialising text hotspots so that they are
able to be displayed immediately. This shouldn't have an major
performance impacts and improves user first impressions.