refactor: status icon hotspot to use images
Use pngs and gifs to display the status as this is significantly simpler
than building the image using ImageDraw.

Update image hotspot to reset the frame number and recreate the cached
image when `image_path` is set. This is so the status icon hotspot is
responsive when it's state is updated. When initialising the image
hotspot the `image_path` property needs to be set after other properties
to prevent "No attribute" exceptions.

Update the test images to include `status/processing.gif` with a single
frame to improve test reliability.