Os 1255 add unit tests (#236)
* chore: remove unused invert property from hotspot

* refactor: replace assistant with font utils

Add `get_font` and `get_mono_font` helpers to utils. The only remaining
area the MiniscreenAssistant is used is to get the recommended font. We
want control over the default font in the application layer rather than
in a dependency. Removing the assistant completely also reduces
complexity.

* test: create initial test environment

Add tests folder with:
- requirements.txt for packages needed to run the tests.
- mocks folder for packages that need specific implementations
- Pitop class mock that allows the user to press miniscreen buttons
- fonts for cross-platform consistency
- images to be used in place of gifs that would cause flaky tests

Use pytest-xdist plugin to run the tests concurrently in different
processes. Splitting the execution across multiple workers significantly
reduces the time to run the tests.

Use pytest-snapshot since that is the only snapshot library that
supports xdist.

Create snapshot reporter plugin that uses imgcat to show failing
snapshots in the terminal as a post-run report.

Disable bootsplash, dimming and screensaver by default to prevent
unexpected test failures.

Modify marquee text hotspot so that it does not scroll during tests.
When testing marquee text hotspot for scrolling behaviour this can be
updated to specify the scroll position.

Add tests for miniscreen mock.

* test: initial snapshots of application

Add initial set of tests to verify pages render correctly in the happy
path.

Add overview page tests
Add hud navigation tests
Add system menu tests
Add network menu tests
Add settings menu tests

* fix: overview starting with no capacity text

The overview page takes about a second to render capacity text next to
the battery. This seems to be because the text hotspot doesn't
initialise it's cached image and the initial render of the page happens
before it's first render cycle.

Render the cached image when initialising text hotspots so that they are
able to be displayed immediately. This shouldn't have an major
performance impacts and improves user first impressions.

* fix: cleanup thread and subscribers when app stopped

* fix: overview page crashing on cleanup

Remove reference to event that is not set `_ip_update_stop_event`.

* refactor: status icon hotspot to use images

Use pngs and gifs to display the status as this is significantly simpler
than building the image using ImageDraw.

Update image hotspot to reset the frame number and recreate the cached
image when `image_path` is set. This is so the status icon hotspot is
responsive when it's state is updated. When initialising the image
hotspot the `image_path` property needs to be set after other properties
to prevent "No attribute" exceptions.

Update the test images to include `status/processing.gif` with a single
frame to improve test reliability.

* build: run tests on PRs and master

Upload coverage reports to Codecov.

* chore: add .virtualenv to gitignore