refactor: use new core API in application
Replace existing application logic with components from the new core
API.

This comes with horizontal transitions for moving into and out of menus
since that is built into the new Stack component.

Some work still needs to be done for this to reach feature-parity:
- screen dimming and screensaver behaviour
- timing out actions
- saving app output to a folder and displaying using imgcat

The base component for text uses more generic "align" and
"vertical_align" properties to layout in it's image. This means some
pages have differences of a pixel or so in their text layout. The
snapshots for these cases have been updated as it is not perceptible to
the eye.

Update pytest opts to ask xdist to group the tests by file so tests in
the same file cannot interfere with one another.