feat: add new core API for miniscreen apps (#244)
The existing application some has aspects that could be improved.

There are many abstractions that have similar responsibilities:

- The majority of the abstractions in the app are reponsible for whether
  they should render, how to render, how to layout their children and
  notifying a parent of a change.

- The number of abstractions make understanding the app difficult as you
  need to learn the place and purpose of each abstraction before it's
  possible to reason about the apps behaviour. This can be confusing
  when abstractions only differ in subtle ways.

- The rigid heirarchy that arises from the structure also makes
  composition difficult, a Tile cannot easily render another Tile for
  example.

Hotspot caching loop inhibits fine-grained control and over-renders.

- Since every change to a hotspot waits for the interval to update the
  cached image it's not easy to make synchronous updates without
  manually updating the cached image directly. This makes fine-grained
  updates difficult / confusing to implement, since creating the image
  is the responsibility of the hotspot in the first place.

- Every hotspot has an interval so even when the render result is
  guaranteed to be unchanged the hotspot's render method is still called.

Add a new core API to resolve the above issues:

Provide App and Component base classes. App takes a root component and
handles starting, stopping, displaying images and setting up the
miniscreen. The Component class is the main abstraction, components:

- define a render method, which takes an image and returns an image.
- define the state needed to render an image
- optionally create child components to be used in the render method.
- optionally create intervals to update state regularly.
- notify parents of a rerender due to a state change or a child rerender

Allowing components to own child components allows us to build a tree of
components. This makes the heirarchy of the app flexible as every layer
is composable with every other layer. Having a single abstraction handle
layout and rendering makes the app easier to understand as well.

When a component calls `self.state.update` it reconciles, which is when
it checks if the output from render has changed and then notifies it's
parent of the rerender if needed. When a child component rerenders the
parent also reconciles itself, this propogates up the tree until either
a parent's output is unchanged or the app displays the resulting image.

To prevent concurrent state updates from causing unexpected behaviour
there is a reconciliation lock per component. This means a parent only
handles a single state update or child rerender at a time. It combines
updates that occur during a reconciliation so that only one subsequent
reconciliation is required to be fired even when many updates happen at
once. This works well for cases when there are a small to medium number
of threads and children per component which is expected to be true in
the majority of cases.

The render method is memoised by default: if it is invoked with the same
image it will return the cached output. This means parents can call a
child's render method frequently without overhead assuming the input
image is unchanged. Combined with the fact that rerenders are triggered
by state changes static components such as text or images have very
little overhead.

Creating intervals to update state within a component was added to
allow for concurrency without exposing the user to full threading.
Intervals created this way are also automatically cleaned up and prevent
circular references and therefore memory leaks being created.

Common components have been added to the components folder. These
components include equivalents of the Text, MarqueeText and Image
hotspots and adds new List, PageList, and Stack components.

Utils for rendering, positioning, fonts and performing timed transitions
have been added to help with common and/or difficult use-cases.

The core is also built with the intention of exposing it through the SDK
eventually. The API is intentionally minimal, only keeping methods that
are essential or handle common use-cases and gotchas.