feat: use new core in hardware and software pages
Replace the apt Cache class with a function that spawns a subprocess in
order to prevent blocking the render thread. For some reason
instantiating the Cache class caused major blocking even when run in a
background thread.

Update the implementation of the hardware and software pages to store
the last results in the class, and where appropriate to only fetch the
info they need once.