feat: allow choosing if List is virtual or not (#258)
Add an argument "virtual" to the List component. When virtual is `True`
the List behaves like it did prior, creating and remove rows as they are
scrolled into and out of the view. When virtual is `False` the List
creates all the rows on creation and keeps them alive even when not
visible.

Default to virtual being False. While virtual lists are better for
creation, processing and memory they have one major drawback: the rows
do not retain their state when scrolled out of view. Since we want the
list to be as simple as possible to use we should default to keeping the
row state around. If the normal list causes performance problems they
can switch to a virtual list which supports arbitrarily large lists.