feat: remove get_text_lazily from Text component (#259)
The Text component calls `get_text` to set the initial text if it is
passed. This means that if `get_text` is slow the main thread becomes
blocked and the app locks up. In order to avoid this `get_text_lazily`
was added to Text: it prevents `get_text` from being called during
initialisation when True.

Instead of having `get_text_lazily` we can stop using `get_text` to set
the initial text entirely. Doing this doesn't actually reduce the
functionality of the Text component since if `text` and `get_text` are
both passed then `text` sets the initial text and `get_text` updates the
text in it's interval.