feat: allow updates to Component state during init
Currently updating Component state at any time triggers a
reconciliation. If state is updated during or soon after initialisation
the reconciliation can cause errors since not all components in the
parent are guaranteed to have been created, or it might conflict with
the initial render that sets up the render cache.

To prevent reconciliation issues during creation don't reconcile until
the Component has been rendered for the first time. This also aligns
better with what reconciliation is meant to be: checking if the render
output has changed and to notify the parent if so.

Add a boolean attribute `rendered` to the Component class to allow
Components to know if the parent has rendered them. This is useful
generally as before being rendered it doesn't make sense to animate
something, and after it is rendered it has a size.

The reason reconciliation is fired when the component has not been
rendered yet is to handle the case where the parent relies on the
child's state to conditionally render the child. This was done to be
tolerant to learners who may not have a good idea of how to organise
state yet. However on closer inspection it only works some of the time,
after initial render the parent is no longer updated when the child
updates.

For now remove the ability of parents to be notified by child component
updates unless the child component has changed it's output. There maybe
a better way to handle this that is more tolerant but for now it is the
most consistent behaviour.