feat: allow waiting for images to finish animating
This is useful in cases where you want to show an image for it's
animation and then hide it again.

Expose the "stop animating" event in the Image component so that parents
can use it to wait for the image to finish animating.

Also set the event when the image finishes animating. This only happens
when seeking over the end of the image frames, so looping images will
never finish.