feat: pause components when not rendered
Intervals running in the background that do not change what is being
displayed on the screen cause performance to drop as an application
scales.

Pause components when they are not being rendered. Pausing a component
means that it's intervals will wait until the component becomes active
again to continue. We know that a component can be paused when it is not
being rendered, since it will never change the output of it's parent.

To track whether a component is being rendered we need to know if it's
render method is called by it's parent's render method. To do this we
can have components mark themselves as rendered whenever their render
method is called. Parents can use this by marking their children as not
rendered before it's render method is called and then check which
children are marked as rendered afterwards. To do this tracking of
whether a component is rendered use the existing `rendered` property.

After rendering or reconciling, a parent needs to set the active state
of it's children based off whether they are rendered or not.

To replace the previous usage of the `rendered` property a new `mounted`
property has been added, this is used to verify reconciliation should
happen and that the render cache is populated.

In order to allow custom threads to hook into this pausing behaviour the
`active_event` that is used to pause intervals has been exposed. One way
to pause a custom thread using this event is to wait for it to become
active by calling `self.active_event.wait`.