perf: pause list rows that are not visible
In order for list rows that are not visible to be paused they must not
be rendered. Currently when the list is not virtual we render all the
rows to then crop the window. Instead of rendering all rows we can
instead determine the exact rows that are needed and only render those
before cropping the window.