test: fix running app suites independently
When only running tests for the app such as `tests/system_test.py` the
setup is broken. It seems like these test rely on a core test suite to
be run as well, presumably because the `setup` fixture needs to be run
before a test that uses the `app` fixture.

Ensure the `app` fixture performs all necessary setup itself so the app
suites can be run independently of the core suites.