fix: restart marquee scrolling when text changes
The carousel used in marquee text to calculate scroll offsets is not
being recreated when text changes, this means that the scroll length can
be wrong after a state update.

Since it's now possible to update the marquee text component before it's
mounted self.width can be None, add checks that self.width is not None
before using it.