fix: app crashing when miniscreen has broken pipe
When some actions are triggered the SPI address of the OLED can change,
this causes a `BrokenPipeError` to be raised when
`miniscreen.device.display` is called. In this case, we should exit the
app with an error, leaving systemd to restart it.

Use app.wait_for_stop instead of signal.pause, so that exiting can be
triggered by a custom event and an error raised.

Update `app.stop` to also clear timers so that they do not remain after
app root has been removed.