fix: service restarting if miniscreen unavailable
Currently when the pt-miniscreen service is run on a device that does
not have the miniscreen it will restart over and over again since it is
set to restart on failure.

Use `ExecCondition` in the systemd service to avoid trying to run the
service on the pi-top [3] or pi-top CEED which we know do not have a
miniscreen. This is more tolerant than checking for if we are running on
a pi-top [4] since we don't know if the service can / should work on
unknown devices.

Another option considered was not running the service if the miniscreen
cannot be instantiated or accessed, but this came with downsides:
- The SDK checks if the device is a pi-top [4] under the hood anyways.
- Importing the SDK takes a long time so delays starting the service.
- On a pi-top [4] we want to restart the service if accessing the
miniscreen fails.