refactor: remove miniscreen from core app
Update the core App class to accept a display method and a size rather
than requiring the Miniscreen to be passed. By removing the dependency
on the Miniscreen class we make the core library useful for any device
that is able to display PIL images. Allow setting the image mode to
support non-monochrome displays.