fix: sped up menu gif after scroll
When scrolling through the Root pages the first frame of the gifs don't
wait for the full duration before changing, giving the appearance that
they are sped up initially.

This is due to the default behaviour of the core List component: it uses
a snapshot of the scrolling rows to improve performance for complex
rows, this snapshot remains on the first frame longer than the gif does,
giving the illusion of it being sped up once scrolling finishes.

Turn off this behaviour of the List by setting
`use_snapshot_while_scrolling` to `False` so that the gif animates while
scrolling, keeping frame time consistent.