feat: add option to not animate List scrolling
It is currently impossible to scroll List components without animating
the scroll.

Add an `animate` argument to methods that scroll the list on the List,
PageList and SelectableList components, for example the `scroll_up` and
`scroll_down` methods on List and PageList or the `select_row` method on
the SelectableList. Note that `select_row` and it's derivatives use the
argument `animate_scroll` to destinguish that it relates to the List
scrolling.

When `animate` is `False` List component scroll methods do not use any
threading / are synchronous.